+++ src/library/inputs/keyboard_helper.h
@@ -45,6 +45,9 @@ void xkeyboardToSDL1keyboard(const std::array<unsigned int,AllInputs::MAXKEYS>& 
 void xkeysymToSDL(SDL_Keysym *keysym, unsigned int xkeysym);
 void xkeysymToSDL1(SDL1::SDL_keysym *keysym, unsigned int xkeysym);
 
+/* Build the SDL2 modifier from the keyboard input */
+SDL_Keymod xkeyboardToSDLMod(const std::array<unsigned int,AllInputs::MAXKEYS>& Xkeyboard);
+
 }
 
 #endif